Add specific search methods for tokens and text.
This is the beginning of a multi-step plan to breakup the monolithic
Search.Search method. There will be a specific search method for each
result type that is accessible though Search.Search, and as a result Search.Search is now deprecated and will be removed when it is no longer needed. There is also a specific options type for each Search method, but they may be unified in the future if appropriate.