race: Synchronise access to NewClientFromContext
If tests which mock `NewClientFromContext` run in parallel, then they make
concurrent writes and reads to `NewClientFromContext` for the purpose of
mocking. Synchronising access doesn't fix the underlying race condition, but in
practice I have not seen it cause flaky test results (yet) and the task of
redesigning this API/fixing the tests is a _lot_ more work for the value gained.